clean up code, removed unused code
