+++ src/components/navbar/links.js
@@ -1,0 +1,21 @@
+const pages = [
+  {
+    page: "About",
+    route: "/#About",
+  },
+
+  {
+    page: "Testimonials",
+    route: "/Testimonials",
+  },
+  {
+    page: "Contact",
+    route: "/#Contact",
+  },
+  {
+    page: "Instagram",
+    route: "https://www.instagram.com/rhythmmessengers/",
+  },
+];
+
+export default pages;
